claude-reflect/pyproject.toml
[project]
name = "claude-reflect"
version = "0.1.0"
description = "MCP server for Claude reflection with memory decay"
# readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Claude Reflect Team" }
]
dependencies = [
    "fastmcp>=0.0.7",
    "qdrant-client>=1.7.0",
    "voyageai>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
claude-reflect = "claude_reflect.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"